+++ src/main/java/site/travellaboratory/be/domain/userlikereview/UserLikeReviewStatus.java
@@ -1,4 +1,4 @@
-package site.travellaboratory.be.domain.userslikerereview;
+package site.travellaboratory.be.domain.userlikereview;
 
 public enum UserLikeReviewStatus {
     ACTIVE,
